pyproject.toml
[project]
name = "ai-foundry-config-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.0",
    "ipykernel>=6.30.1",
    "pip>=25.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[tool.uv.workspace]
members = [
    "agents",
]
exclude = [
    "agents_v2",
    "subscription-manager",
    "options-infra/foundry-search/scripts",
    "options-infra/foundry-rag-search/scripts",
    "options-infra/foundry-search/scripts/agents_v2",
    "options-infra/foundry-rag-search/scripts/agents_v2",
    "options-infra/foundry-search/scripts/agents_v1",
    "options-infra/foundry-rag-search/scripts/agents_v1",
    "options-infra/foundry-sk-agents/apps/sk-agents"
]

